refactor(control-plane): move status runtime summaries into status bounded context
