+++ loopx/control_plane/status/runtime_summaries.py
@@ -1,10 +1,12 @@
+"""Status runtime summaries inside the `status` bounded context."""
+
 from __future__ import annotations
 
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Callable, Iterable
 
-from .runtime.decision_freshness import (
+from ..runtime.decision_freshness import (
     DECISION_FRESHNESS_CLASSIFICATION_PREFIXES,
     DECISION_FRESHNESS_ITEM_LIMIT,
     DECISION_FRESHNESS_PROXY_NOTE,
@@ -12,15 +14,15 @@ from .runtime.decision_freshness import (
     build_decision_freshness_summary,
     decision_event_kinds as _decision_event_kinds_read_model,
 )
-from .runtime.event_ledger import (
+from ..runtime.event_ledger import (
     blank_event_class_counts,
     build_event_ledger_summary,
     event_ledger_event_class as _event_ledger_event_class_read_model,
 )
-from .runtime.promotion_readiness import build_promotion_readiness_summary
-from .runtime.run_history import build_run_history
-from .quota.usage_summary import build_usage_summary
-from .todos.todo_index import build_todo_index
+from ..runtime.promotion_readiness import build_promotion_readiness_summary
+from ..runtime.run_history import build_run_history
+from ..quota.usage_summary import build_usage_summary
+from ..todos.todo_index import build_todo_index
 
 
 StatusCallback = Callable[..., Any]
